Add comments in code and a bit refactor
