+++ srcs/server/PollFds.hpp
@@ -7,6 +7,10 @@
 
 namespace server {
 	class PollFds {
+		public:
+			struct PollException : public std::exception {
+				virtual const char* what() const throw();
+			};
 		private:
 			// number of listeners socket
 			int _listeners_num;
@@ -25,17 +29,14 @@ namespace server {
 			// close of all sockets(file descriptors)
 			~PollFds();
 
-			// give data for poll syscall (first argument in poll)
-			struct pollfd* get_array();
-			// give data for poll syscall (second argument in poll)
-			int            get_array_size();
-
+			// do_poll poll syscall wrapper, throw exception in case of error
+			void do_poll();
 			// add new listener in class
-			void           add_listener(int listener);
+			void add_listener(int listener);
 			// add new client in class
-			void           add_client(int sock);
+			void add_client(int sock);
 			// erase sock from class and close it
-			void           erase_sock(int sock);
+			void erase_sock(int sock);
 
 			// check actions on termination socket after poll call
 			const std::set<int>& check_term();
